requirements.txt
pylint
requests
unidecode
pillow
